displacement

+++ src/main/java/ch/geomo/tramaps/map/displacement/DisplaceGuard.java
@@ -28,7 +28,7 @@ public class DisplaceGuard {
         traversedNodes = new ArrayList<>();
     }
 
-    public double getMoveDistance() {
+    public int getMoveDistance() {
         return conflict.getBestDisplacementLength();
     }
